Khant.X
Hello,nice to meet you everyone in this.
I am from Myanmar(burma).
I just wanna to learn about something like a system docs that how it  is work?
It work for what? When time its will start working just like will this.
Please help me to leaning the about of system working and when time it  decease some action.